+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
--
+- [fixed] Fixed optional fields in UserRecord types to be optional.
 
 # v6.4.0